コンポーネント内のfetchリクエスト処理を共通化
