+++ app/javascript/components/OtherForm.jsx
@@ -1,5 +1,6 @@
 import { useState } from 'react'
 import PropTypes from 'prop-types'
+import sendRequest from '../sendRequest.js'
 
 export default function OtherForm({ minuteId, content }) {
   const [inputValue, setInputValue] = useState(content)
@@ -11,18 +12,12 @@ export default function OtherForm({ minuteId, content }) {
   const handleClick = async function (e) {
     e.preventDefault()
     const parameter = { minute: { other: inputValue } }
-    const csrfToken = document.head.querySelector(
-      'meta[name=csrf-token]'
-    )?.content
 
-    const response = await fetch(`/api/minutes/${minuteId}`, {
-      method: 'PATCH',
-      body: JSON.stringify(parameter),
-      headers: {
-        'Content-Type': 'application/json; charset=utf-8',
-        'X-CSRF-Token': csrfToken,
-      },
-    })
+    const response = await sendRequest(
+      `/api/minutes/${minuteId}`,
+      'PATCH',
+      parameter
+    )
 
     if (response.status !== 200) {
       const errorData = await response.json()
